fourteen guards that could not fire: twelve un-darkened, and the detector ratcheted so a thirteenth is a CI failure
`cargo clippy --all-targets` ABORTS at the first non-compiling target, so nobody
running it locally saw past `dregg-turn`. With `--keep-going` the workspace shows
FOURTEEN dark targets, not the nine that were visible. Two of the extra five were
HIDDEN BEHIND A RED LIB: `starbridge-v2`'s lib has not compiled since 2026-07-17,
and a crate whose lib is red never gets its own test targets attempted — fixing the
lib is what made `deos_desktop_editor_types_to_heap` and `deos_doc_persists_to_cell_heap`
visible at all. A count of dark targets taken while a lib is red is a LOWER BOUND.

STALE TEST vs BROKEN CODE, per target. Every verdict below was taken at the
definition and against git history, never from a commit subject.

SIX were STALE CALL SITES — a deliberate, well-motivated change that swept its
siblings and left one caller behind:
  * `sdk/tests/sovereign_rotated_wide.rs` — abe4a8c923 flipped the wide refusal
    route off the lossy one-felt leaf (`fields_root_leaves` + `fold_bytes32`) onto
    the exact 32-raw-byte leaf, updating 3 of 4 call sites.
  * `starbridge-v2/src/app_registry.rs` (PRODUCTION, default features, red 8 days) —
    a13f582de6 added the `ballots` parameter that binds a tally to an exhibited
    ballot set. See below; this one had a trap in it.
  * `starbridge-web-surface` src+tests (5 sites) — 37a190357d widened
    `Effect::SetField.index` usize→u64 because a `usize` key made one SIGNED turn
    mean different things on wasm32 and native. A real wire-format fix.
  * `circuit-prove/tests/gnark_witness_export_teeth.rs` — 91708cd5fb added
    `board_window`; its bridge sibling was repaired on 07-25, this one was the last
    unrepaired literal construction.
  * `starbridge-v2/tests/deos_{desktop_editor_types_to_heap,doc_persists_to_cell_heap}.rs`
    — `CellState::heap_root` is now the wide `Faithful8`, the felt-width type wall.
  * `teasting/tests/predicate_soundness.rs` — 1cdc7fe66b made
    `BridgePredicateProof::blinding` an `Option` (`None` = the third-party attested
    shape, which discloses no opening). The file was swept in mid-flight by
    cf84a9baa7 ten hours AFTER the API moved, so it has never compiled.

FOUR NEVER COMPILED ONCE, and three of those were born red in the commit that
introduced them:
  * `turn/tests/umem_v2_faithful_codec.rs` — `UKey::Cell` has ZERO occurrences in
    all of history outside the commit that wrote it. Added by 5b7173f0f0; see below.
  * `circuit/examples/bfv_stage_canary.rs` + `circuit/src/private_book_bfv_exact_public.rs`
    — added together by abe4a8c923 with `#[derive(Clone, Debug)]` on a struct holding
    a `p3_batch_stark::BatchProof`, which has neither at the pinned rev, and a
    `Result<_, String>` signature over a function that had already stopped returning
    `String`.
  * `exec-lean/tests/lean_state_producer_coverage.rs` — 5c5feb1761 changed
    `let (pre, a_id)` to `let (_, a_id)` and left the consumer 13 lines below. The
    SAME commit added three accumulator-fence teeth. They have never executed.
  * `circuit-prove/tests/private_shuffle_fair_zk.rs` — DELETED, not repaired. It was
    a two-line `#[path = "../src/…"]` shim from before the module was promoted to
    `pub mod` in lib.rs; under the shim `crate::` is the test binary, which has no
    `private_shuffle`. The four tests it was reaching run under `--lib` today. Both
    such shims in the tree were red; this pattern had a 100% failure rate. (The
    deletion is not in THIS commit: a co-tenant lane's `60932256ae` swept the staged
    removal up. Same outcome, different author line.)

TWO STAY DARK and are enumerated instead: the FRI soundness ledger gate. It was
killed in TWO stages, 17 hours apart, by two commits NEITHER of which mentions FRI —
7b70d6a693 ("no-silent-fallback … PQ") dropped build.rs's `dregg_fri_ledger` probe
and both its cfg emissions, then a8f9a82368 ("ML-KEM keygen …") deleted the Rust API
and lean_init.c's C shim. Stage 1 alone had already stopped it; stage 2 turned a loud
failure into an unresolved import. The repair is confined to `dregg-lean-ffi/`, a live
lane's territory, so it is written out in full in `.github/dark-targets.txt`. The Lean
half is alive and orphaned (`@[export] dregg_fri_ledger`, inside the FFI boundary
closure, so the symbol IS in today's archive). MEASURED rather than assumed: no knob
drifted in the five days it was dark — all 7 shipped configs still equal their Lean
literals.

THE 5b7173f0f0 VERDICT ("17 encoder bodies COLLAPSED or DELETED, zero bytes moved").
The subject is not refuted; it is UNTESTED, by its own hand. That commit added
`umem_v2_faithful_codec.rs` — the pin that would have caught a byte move on the one
surviving endianness divergence — and the file names a `UKey` variant that does not
exist and never has. It could not have compiled when it was written. The 8 red
`dregg-persist` lib tests are NOT its doing: `5b7173f0f0` touches no file under
`persist/`, and the receipt encoder persist actually calls
(`exact_nullifier_aafi::{Digest8, exact_state_commit}`) is not in its 43-file diff.
Separately: no CI workflow runs `dregg-persist` tests at all, so those 8 have never
been gated. Their cause is still open and belongs to a persist lane.

TWO FINDINGS THE REPAIRS TURNED UP, worth more than the compile fixes:
  * `sovereign_rotated_wide.rs` hand-pinned TWO wide-descriptor PI bases, and BOTH
    were stale — measured off the staged registry at HEAD, `transferVmDescriptor2R24`
    is 68 PIs (base 52; the file said 46) and `refusalVmDescriptor2R24` is 90 (base
    74; the file said 54). A forgery tooth writing six lanes early perturbs a PI that
    is not a commit felt at all: it would have gone on "biting" the wrong thing. The
    bases are now DERIVED (`public_input_count - 16`, the sibling wide tests' idiom);
    only the arity stays pinned, as a canary that must be re-measured.
  * The obvious green-making repair for `app_registry.rs` is the WRONG one. Passing
    `&BTreeSet::new()` fails closed (`CountGe` requires ≥1) — a dead button. Passing a
    fabricated id COMMITS, because `CountGe` does not check element provenance — which
    re-launders exactly the tally forgery a13f582de6 closed. So the seed closure now
    BIRTHS the companion ballot (`seed_ballot`) and the fire exhibits that real cell.

THE DETECTOR. `.github/dark-targets.txt` + a `dark-target ratchet` step on the
existing `clippy-correctness` job (6154a0cb40's), which already sweeps every target
with `--keep-going`. That job REPORTED non-compiling targets as a `::warning::`
annotation — a documented wound, not a detected one. Now: a target that goes dark and
is not on the list FAILS the job, and a row that compiles again FAILS it too (a stale
allowlist is how the next one gets waved through). The list carries a reason and an
owner per row, and today it has exactly two rows.

Verified: `cargo check --all-targets --keep-going` over all eight touched packages —
`dregg-{sdk,teasting,exec-lean,turn,circuit,circuit-prove}`, `starbridge-{v2,web-surface}`
— reports the two FRI targets and nothing else.

+++ starbridge-web-surface/src/affordance.rs
@@ -1613,7 +1613,7 @@ mod tests {
             intent.effect_summary(),
             EffectSummary::SetField {
                 cell: doc,
-                index: TALLY_SLOT
+                index: TALLY_SLOT as u64
             }
         );
         assert!(matches!(intent.effect, Effect::SetField { .. }));
